+++ crates/connectors/ndc-postgres/src/query.rs
@@ -41,6 +41,8 @@ pub async fn query(
             query_request = ?query_request
         );
 
+        let request_arguments = query_request.request_arguments.clone();
+
         let plan = async {
             plan_query(configuration, state, query_request).map_err(|err| {
                 record::translation_error(&err, &state.query_metrics);
@@ -50,7 +52,14 @@ pub async fn query(
         .instrument(info_span!("Plan query"))
         .await?;
 
-        let pool = state.pool_manager.acquire();
+        let pool = state
+            .pool_manager
+            .acquire(&request_arguments, &state.query_metrics)
+            .await
+            .map_err(|err| {
+                record::pool_acquisition_error(&err, &state.query_metrics);
+                convert::pool_acquisition_error_to_response(&err)
+            })?;
 
         let result = async {
             execute_query(&pool.pool, &pool.database_info, &state.query_metrics, plan)