+++ crates/connectors/ndc-postgres/src/state.rs
@@ -4,15 +4,21 @@
 
 use ndc_postgres_configuration::PoolSettings;
 use ndc_postgres_configuration::{get_connect_options, ConnectionSettings, Redacted, SslInfo};
+use ndc_sdk::models::ArgumentName;
 use percent_encoding::percent_decode_str;
 use query_engine_execution::database_info::{self, DatabaseInfo, DatabaseVersion};
 use query_engine_execution::metrics;
 use sqlx::postgres::{PgPool, PgPoolOptions, PgRow};
 use sqlx::{Connection, Row};
-use std::sync::Arc;
+use std::collections::BTreeMap;
+use std::sync::atomic::{AtomicUsize, Ordering};
+use std::sync::{Arc, RwLock};
 use thiserror::Error;
 use tracing::{info_span, Instrument};
 use url::Url;
+
+const CONNECTION_NAME_ARGUMENT: &str = "connection_name";
+const CONNECTION_STRING_ARGUMENT: &str = "connection_string";
 
 /// State for our connector.
 #[derive(Debug)]
@@ -21,6 +27,9 @@ pub struct State {
     pub query_metrics: metrics::Metrics,
     pub configuration_metrics: ndc_postgres_configuration::Metrics,
 }
+
+type DatabaseConnectionName = String;
+type DatabaseConnectionString = String;
 
 /// A connection pool bundled with database info and a unique index
 #[derive(Debug)]
@@ -80,7 +89,24 @@ impl PoolContext {
 /// A pool manager that possibly handles multiple connection pools
 #[derive(Debug)]
 pub enum PoolManager {
-    Static { pool: Arc<PoolContext> },
+    Static {
+        pool: Arc<PoolContext>,
+    },
+    Named {
+        fallback_pool: Option<Arc<PoolContext>>,
+        pools: Arc<RwLock<BTreeMap<DatabaseConnectionName, Arc<PoolContext>>>>,
+        connection_uris: BTreeMap<DatabaseConnectionName, Redacted<DatabaseConnectionString>>,
+        ssl: Redacted<SslInfo>,
+        pool_settings: PoolSettings,
+        next_pool_index: AtomicUsize,
+    },
+    Dynamic {
+        fallback_pool: Option<Arc<PoolContext>>,
+        pools: Arc<RwLock<BTreeMap<Redacted<DatabaseConnectionString>, Arc<PoolContext>>>>,
+        ssl: Redacted<SslInfo>,
+        pool_settings: PoolSettings,
+        next_pool_index: AtomicUsize,
+    },
 }
 
 impl PoolManager {
@@ -96,6 +122,75 @@ impl PoolManager {
             } => Ok(Self::Static {
                 pool: Arc::new(PoolContext::new(connection_uri, ssl, pool_settings, 0).await?),
             }),
+            ConnectionSettings::Named {
+                fallback_connection_uri,
+                fallback_to_static,
+                ssl,
+                connection_uris,
+                eager_connections,
+            } => {
+                let next_pool_index = AtomicUsize::new(0);
+
+                let fallback_pool = if *fallback_to_static {
+                    let index = next_pool_index.fetch_add(1, Ordering::SeqCst);
+
+                    Some(Arc::new(
+                        PoolContext::new(fallback_connection_uri, ssl, pool_settings, index)
+                            .await?,
+                    ))
+                } else {
+                    None
+                };
+
+                let mut pools = BTreeMap::new();
+                if *eager_connections {
+                    for (connection_name, connection_uri) in connection_uris {
+                        let index = next_pool_index.fetch_add(1, Ordering::SeqCst);
+
+                        pools.insert(
+                            connection_name.clone(),
+                            Arc::new(
+                                PoolContext::new(connection_uri, ssl, pool_settings, index).await?,
+                            ),
+                        );
+                    }
+                }
+
+                Ok(Self::Named {
+                    fallback_pool,
+                    pools: Arc::new(RwLock::new(pools)),
+                    connection_uris: connection_uris.clone(),
+                    ssl: ssl.clone(),
+                    pool_settings: pool_settings.clone(),
+                    next_pool_index,
+                })
+            }
+            ConnectionSettings::Dynamic {
+                fallback_connection_uri,
+                fallback_to_static,
+                ssl,
+            } => {
+                let next_pool_index = AtomicUsize::new(0);
+
+                let fallback_pool = if *fallback_to_static {
+                    let index = next_pool_index.fetch_add(1, Ordering::SeqCst);
+
+                    Some(Arc::new(
+                        PoolContext::new(fallback_connection_uri, ssl, pool_settings, index)
+                            .await?,
+                    ))
+                } else {
+                    None
+                };
+
+                Ok(Self::Dynamic {
+                    fallback_pool,
+                    pools: Arc::new(RwLock::new(BTreeMap::new())),
+                    ssl: ssl.clone(),
+                    pool_settings: pool_settings.clone(),
+                    next_pool_index,
+                })
+            }
         }
     }
 
@@ -124,33 +219,240 @@ impl PoolManager {
     ///
     /// If the request does not include a connection string, check if we have a fallback pool, which we will if fallbackToStatic is set.
     /// If we do, return the fallback pool. Else, error.
-    pub fn acquire(&self) -> Arc<PoolContext> {
+    pub async fn acquire(
+        &self,
+        request_arguments: &Option<BTreeMap<ArgumentName, serde_json::Value>>,
+        metrics: &metrics::Metrics,
+    ) -> Result<Arc<PoolContext>, PoolAcquisitionError> {
         match self {
-            Self::Static { pool } => pool.clone(),
-        }
-    }
+            Self::Static { pool } => Ok(pool.clone()),
+            Self::Named {
+                fallback_pool,
+                connection_uris,
+                pools,
+                ssl,
+                pool_settings,
+                next_pool_index,
+            } => {
+                // Extract the connection name from the request arguments
+                if let Some(connection_name) = request_arguments
+                    .as_ref()
+                    .and_then(|request_arguments| request_arguments.get(CONNECTION_NAME_ARGUMENT))
+                    .and_then(|connection_name| connection_name.as_str())
+                {
+                    {
+                        let read_guard = pools.read().map_err(LockError::from)?;
+                        if let Some(pool) = read_guard.get(connection_name) {
+                            return Ok(pool.clone());
+                        }
+                    }
 
-    /// Update the pool options metrics for all pools
-    pub fn set_pool_options_metrics_all(&self, metrics: &metrics::Metrics) {
-        match self {
-            Self::Static { pool } => {
-                metrics.set_pool_options_metrics(pool.pool.options(), pool.index, "default");
+                    if let Some(connection_uri) = connection_uris.get(connection_name) {
+                        // Get a new pool index. If the pool creation fails, we may not use that index ever, but that's fine, we only care that indexes are unique.
+                        let index = next_pool_index.fetch_add(1, Ordering::SeqCst);
+
+                        // Create a new pool for this connection
+                        let pool_instance = Arc::new(
+                            PoolContext::new(connection_uri, ssl, pool_settings, index).await?,
+                        );
+
+                        {
+                            // Get a write lock
+                            let mut write_guard = pools.write().map_err(LockError::from)?;
+
+                            // Check again to make sure someone else didn't create it
+                            if let Some(pool) = write_guard.get(connection_name) {
+                                return Ok(pool.clone());
+                            }
+
+                            // Insert it into the map
+                            write_guard.insert(connection_name.to_string(), pool_instance.clone());
+                        }
+
+                        // update pool options metrics as we've added a new pool
+                        self.set_pool_options_metrics_all(metrics)?;
+
+                        // Return it
+                        Ok(pool_instance)
+                    } else {
+                        // No connection URI found for this name
+                        Err(PoolAcquisitionError::UnknownConnectionName(
+                            connection_name.to_string(),
+                        ))
+                    }
+                } else if let Some(fallback_pool) = fallback_pool {
+                    Ok(fallback_pool.clone())
+                } else {
+                    Err(PoolAcquisitionError::MissingRequiredRequestArgument(
+                        CONNECTION_NAME_ARGUMENT.to_string(),
+                    ))
+                }
+            }
+            Self::Dynamic {
+                fallback_pool,
+                pools,
+                ssl,
+                pool_settings,
+                next_pool_index,
+            } => {
+                // Extract the connection string from the request arguments
+                if let Some(connection_string) = request_arguments
+                    .as_ref()
+                    .and_then(|request_arguments| request_arguments.get(CONNECTION_STRING_ARGUMENT))
+                    .and_then(|connection_string| connection_string.as_str())
+                {
+                    // Create a redacted version of the connection string for use as a key
+                    let redacted_connection_string = Redacted::new(connection_string.to_string());
+
+                    // Check if we already have a pool for this connection string
+                    {
+                        let read_guard = pools.read().map_err(LockError::from)?;
+                        if let Some(pool) = read_guard.get(&redacted_connection_string) {
+                            return Ok(pool.clone());
+                        }
+                    }
+
+                    // Get a new pool index. If the pool creation fails, we may not use that index ever, but that's fine, we only care that indexes are unique.
+                    let index = next_pool_index.fetch_add(1, Ordering::SeqCst);
+
+                    // Create a new pool for this connection
+                    let pool_instance = Arc::new(
+                        PoolContext::new(&redacted_connection_string, ssl, pool_settings, index)
+                            .await?,
+                    );
+
+                    {
+                        // Get a write lock
+                        let mut write_guard = pools.write().map_err(LockError::from)?;
+
+                        // Check again to make sure someone else didn't create it
+                        if let Some(pool) = write_guard.get(&redacted_connection_string) {
+                            return Ok(pool.clone());
+                        }
+
+                        // Insert it into the map
+                        write_guard.insert(redacted_connection_string, pool_instance.clone());
+                    }
+
+                    // update pool options metrics as we've added a new pool
+                    self.set_pool_options_metrics_all(metrics)?;
+
+                    // Return it
+                    Ok(pool_instance)
+                } else if let Some(fallback_pool) = fallback_pool {
+                    Ok(fallback_pool.clone())
+                } else {
+                    Err(PoolAcquisitionError::MissingRequiredRequestArgument(
+                        CONNECTION_STRING_ARGUMENT.to_string(),
+                    ))
+                }
             }
         }
     }
 
-    /// Update the pool metrics for all pools
-    pub fn update_pool_metrics_all(&self, metrics: &metrics::Metrics) {
+    pub fn set_pool_options_metrics_all(
+        &self,
+        metrics: &metrics::Metrics,
+    ) -> Result<(), LockError> {
+        match self {
+            Self::Static { pool } => {
+                metrics.set_pool_options_metrics(pool.pool.options(), pool.index, "default");
+                Ok(())
+            }
+            Self::Named {
+                fallback_pool,
+                pools,
+                ..
+            } => {
+                if let Some(fallback_pool) = fallback_pool {
+                    metrics.set_pool_options_metrics(
+                        fallback_pool.pool.options(),
+                        fallback_pool.index,
+                        "fallback",
+                    );
+                }
+                for (poolname, pool) in pools.read()?.iter() {
+                    metrics.set_pool_options_metrics(pool.pool.options(), pool.index, poolname);
+                }
+                Ok(())
+            }
+            Self::Dynamic {
+                fallback_pool,
+                pools,
+                ..
+            } => {
+                if let Some(fallback_pool) = fallback_pool {
+                    metrics.set_pool_options_metrics(
+                        fallback_pool.pool.options(),
+                        fallback_pool.index,
+                        "fallback",
+                    );
+                }
+                for pool in pools.read()?.values() {
+                    metrics.set_pool_options_metrics(pool.pool.options(), pool.index, "dynamic");
+                }
+                Ok(())
+            }
+        }
+    }
+
+    pub fn update_pool_metrics_all(&self, metrics: &metrics::Metrics) -> Result<(), LockError> {
         match self {
             Self::Static { pool } => {
                 metrics.update_pool_metrics(&pool.pool, pool.index, "default");
+                Ok(())
+            }
+            Self::Named {
+                fallback_pool,
+                pools,
+                ..
+            } => {
+                if let Some(fallback_pool) = fallback_pool {
+                    metrics.update_pool_metrics(
+                        &fallback_pool.pool,
+                        fallback_pool.index,
+                        "fallback",
+                    );
+                }
+                for (poolname, pool) in pools.read()?.iter() {
+                    metrics.update_pool_metrics(&pool.pool, pool.index, poolname);
+                }
+                Ok(())
+            }
+            Self::Dynamic {
+                fallback_pool,
+                pools,
+                ..
+            } => {
+                if let Some(fallback_pool) = fallback_pool {
+                    metrics.update_pool_metrics(
+                        &fallback_pool.pool,
+                        fallback_pool.index,
+                        "fallback",
+                    );
+                }
+                for pool in pools.read()?.values() {
+                    metrics.update_pool_metrics(&pool.pool, pool.index, "dynamic");
+                }
+                Ok(())
             }
         }
     }
 }
 
 #[derive(Debug, thiserror::Error)]
-pub enum PoolAquisitionError {}
+pub enum PoolAcquisitionError {
+    #[error("Missing required request argument: {0}")]
+    MissingRequiredRequestArgument(String),
+    #[error("Invalid value for argument: {0}")]
+    InvalidRequestArgument(String),
+    #[error("Unknown connection name {0}")]
+    UnknownConnectionName(String),
+    #[error("Failed to acquire lock: {0}")]
+    LockError(#[from] LockError),
+    #[error("Failed to create pool: {0}")]
+    PoolCreationError(#[from] InitializationError),
+}
 
 #[derive(Debug, thiserror::Error)]
 pub enum LockError {
@@ -176,7 +478,7 @@ pub async fn create_state(
     let (query_metrics, configuration_metrics) = async {
         let query_metrics_inner = metrics::Metrics::initialize(metrics_registry)
             .map_err(InitializationError::MetricsError)?;
-        pool_manager.set_pool_options_metrics_all(&query_metrics_inner);
+        pool_manager.set_pool_options_metrics_all(&query_metrics_inner)?;
 
         let configuration_metrics_inner =
             ndc_postgres_configuration::Metrics::initialize(metrics_registry)